NOISSUE - Add Prettier Check to CI (#60)
* Fix formatting issues with Prettier and add Prettier check to CI

This commit fixes formatting issues in the codebase by running Prettier and adds a Prettier check to the CI workflow. This ensures consistent code formatting across the project.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* feat(ci): add prettier check for UI folder

This commit adds a new feature to the CI workflow by including a prettier check specifically for the UI folder. Previously, the prettier check was performed on the entire codebase. With this change, the check will only be applied to the UI folder, ensuring consistent code formatting in that specific section of the codebase.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

---------

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>