NOISSUE - Abstract list data-structure on input form (#32)
* add generate color and deleteItem functions

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix bug with list item highlighting and add functionality to add items to a list using the Enter key.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix prettier version in package-lock.json

The package-lock.json file had an incorrect version of prettier. This commit fixes the version to "^3.1.0" to ensure compatibility with the project dependencies.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix styling issue in listUtils.js and domains.html

The commit fixes a styling issue in the listUtils.js file by adding a margin-bottom class to the newly created list item. This ensures proper spacing between list items.

In addition, the commit also addresses a styling issue in the domains.html file. The commit modifies the class attribute of a div element to ensure proper alignment within a row.

These changes improve the overall visual consistency and user experience.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix bug in listUtils.js

The deleteItem function in listUtils.js was not working properly due to unnecessary code. Removed the unnecessary code and fixed the bug.

Also, removed an extra line in bootstraps.html.

This commit addresses the above issues.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix bug in listUtils.js

The commit fixes a bug in the listUtils.js file where an empty input value was being added as a list item. The bug is fixed by adding a condition to check if the input value is not empty before creating a new list item.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix bug in deleteItem function

The deleteItem function in listUtils.js was not correctly checking for the class name "highlight-list-item" when deleting an item. This has been fixed by using the classList.contains method instead of comparing the className directly.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Update ui/web/template/users.html

Co-authored-by: Ian Ngethe Muchiri <100555904+ianmuchyri@users.noreply.github.com>

* Update ui/web/template/things.html

Co-authored-by: Ian Ngethe Muchiri <100555904+ianmuchyri@users.noreply.github.com>

* Refactor listUtils.js to add buttons to list items

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix button class in listUtils.js and add id attribute in navbar.html

The button class in listUtils.js was updated to "btn btn-primary" to match the styling of other buttons. In navbar.html, an id attribute was added to the "domain-tags" input element.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* fix linting

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix button class name in listUtils.js

The commit fixes the class name of the button in listUtils.js.
The previous class name "btn-primary" has been changed to "body-button".
This ensures consistency with the rest of the codebase and improves readability.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Refactor tag identifier naming

Standardize and clarify tag-related variable names in HTML templates for things and users. Renamed 'tags' to 'thingsTags' and 'userTags' to mitigate ambiguity across different features where tag systems are utilized. These changes improve code readability and maintain a consistent naming convention.

Signed-off-by: SammyOina <sammyoina@gmail.com>

---------

Signed-off-by: SammyOina <sammyoina@gmail.com>
Co-authored-by: Ian Ngethe Muchiri <100555904+ianmuchyri@users.noreply.github.com>