NOISSUE - Implement registration page (#119)
* Implement registration page

Implement logging and metrics middleware for ViewRegistration and RegisterUser methods in API. Add functions for request validation, decoding register user requests, handling registration process, and creating login functionality. Include HTML templates for registration and login pages, and JavaScript function for form submission.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* Lint code with prettier

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* Refactored registration form submission logic

In this commit, I refactored the registration form submission logic in the `registration.html` template. I removed the switch statement and added a promise chain to handle the response from the server. Now, the response is always expected to be JSON, and the user's information is stored in the local storage. After successful registration, the user is redirected to the domains page.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* fix(registration): Handle conflict cases with errConflict

This commit modifies error variables and handles error cases in three different files. It adds the "errConflict" error and uses it to handle conflict cases in the registration process. The code sends a request to create a new user, handles the response, and displays error messages or saves user data as needed.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* feat(ui): add click event listener to register button

This commit adds a click event listener to the register button in the registration.html template. When the register button is clicked, the registration error message is cleared.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* bug: Rename showError from  showAlert

The showError function is used to display error messages in different scenarios, such as invalid login credentials or existing user email during registration.
These modifications align with web standards between alert which is inbuilt on the browser

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

---------

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>