NOISSUE - Check for unused params (#126)
* Check for unused params

- Updated logging middleware and metrics middleware
- Removed unused parameters and updated method calls

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* Refactor showUpdatePasswordEndpoint and decodeShowPasswordUpdate functions

The changes in this commit include refactoring the showUpdatePasswordEndpoint and decodeShowPasswordUpdate functions in the ui/api/endpoint.go and ui/api/transport.go files. The showUpdatePasswordEndpoint function now accepts a request parameter of type interface{} and does not perform any validation. The decodeShowPasswordUpdate function no longer retrieves a token from a cookie and returns nil instead.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

---------

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>